requirements.txt

pandas==2.0.3
pillow==10.3.0
python-dotenv==1.0.1
pytesseract == 0.3.10
openai == 1.33.0
fastapi==0.111.0
uvicorn==0.30.1